+++ tests/unit/components/Camera/PhotoCarousel.test.js
@@ -1,13 +1,8 @@
 import { faker } from "@faker-js/faker";
-import {
-  fireEvent, render, screen, waitFor
-} from "@testing-library/react-native";
+import { render, screen } from "@testing-library/react-native";
 import PhotoCarousel from "components/Camera/StandardCamera/PhotoCarousel";
 import initI18next from "i18n/initI18next";
 import React from "react";
-import useStore from "stores/useStore";
-
-const initialStoreState = useStore.getState( );
 
 const mockPhotoUris = [
   faker.image.imageUrl( ),
@@ -17,7 +12,6 @@ const mockPhotoUris = [
 
 describe( "PhotoCarousel", ( ) => {
   beforeAll( async () => {
-    useStore.setState( initialStoreState, true );
     await initI18next( );
   } );
   // There were some tests of photo sizes responding to the isLargeScreen prop
@@ -39,29 +33,5 @@ describe( "PhotoCarousel", ( ) => {
 
     // Snapshot test
     expect( screen ).toMatchSnapshot();
-  } );
-
-  it( "deletes a photo on long press", async ( ) => {
-    useStore.setState( {
-      evidenceToAdd: [mockPhotoUris[2]],
-      cameraPreviewUris: mockPhotoUris,
-      photoEvidenceUris: mockPhotoUris
-    } );
-
-    render(
-      <PhotoCarousel photoUris={mockPhotoUris} isLargeScreen deletePhoto={jest.fn( )} />
-    );
-    const photoImage = screen.getByTestId( `PhotoCarousel.displayPhoto.${mockPhotoUris[2]}` );
-    fireEvent( photoImage, "onLongPress" );
-    const deleteMode = screen.getByTestId( `PhotoCarousel.deletePhoto.${mockPhotoUris[2]}` );
-    await waitFor( ( ) => {
-      expect( deleteMode ).toBeVisible( );
-    } );
-    fireEvent.press( deleteMode );
-
-    const deletedPhoto = screen.queryByTestId( `PhotoCarousel.displayPhoto.${mockPhotoUris[2]}` );
-    await waitFor( ( ) => {
-      expect( deletedPhoto ).toBeFalsy( );
-    } );
   } );
 } );
